Delete method added

+++ lib/complete_me.rb
@@ -1,4 +1,5 @@
 require_relative '../lib/node'
+require 'pry'
 
 class CompleteMe
 
@@ -41,11 +42,13 @@ class CompleteMe
   end
 
   def suggest(fragment)
+    suggestions = []
     node = node_finder(fragment)
     if node!= nil
       second_halves = find_suggestions(node)
       suggestions = second_halves.map {|second_half| fragment + second_half}
     end
+    suggestions
   end
 
   def search(length, letter, characters, node)
@@ -57,15 +60,41 @@ class CompleteMe
   end
 
   def node_finder(fragment, node = @root_node)
+    return @root_node if fragment == ""
     length = fragment.length
     characters = fragment.chars
     letter = characters.delete_at(0)
-    search(length, letter, characters, node) if node.includes_link?(fragment[0])
+    search(length, letter, characters, node) if node.includes_link?(letter)
   end
 
   def populate(file)
     words = File.readlines(file)
     words.each {|word| insert(word.gsub(/\n/, ""))}
+  end
+
+  def vanish_nodes(node, characters, letter)
+    if node != @root_node && node.links.length == 0
+      node.disappear
+      delete_nodes(characters.join, letter)
+    end
+  end
+
+  def delete_nodes(word, previous_letter = "")
+    node = node_finder(word)
+    characters = word.chars
+    letter = characters.delete_at(-1)
+    if node != nil
+      node.delete_key(previous_letter)
+      vanish_nodes(node, characters, letter)
+    end
+  end
+
+  def delete(word)
+    terminator_node = node_finder(word)
+    if terminator_node != nil
+      terminator_node.remove_terminator 
+      delete_nodes(word) if terminator_node.links.length == 0
+    end
   end
 
 end
